Выполненное второе домашнее задание

+++ src/main/java/com/teachmeskill/homework/hometask2/Task2.java
@@ -2,8 +2,7 @@ package com.teachmeskill.homework.hometask2;
 
 public class Task2 {
     public static void main(String[] args) {
-
-        for (long myNumber = 10; myNumber <= 20; myNumber++) {
+        for (int myNumber = 10; myNumber <= 20; myNumber++) {
             System.out.println("Квадрат числа " + myNumber + " = " + myNumber * myNumber);
         }
     }
